Move compileRA function to separate module
